requirements.txt
Django==3.1.7
djangorestframework==3.12.4
factory-boy==3.2.0
Faker==8.1.0
